+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - `[jest-haste-map]` Add missing `@types/graceful-fs` dependency ([#9913](https://github.com/facebook/jest/pull/9913))
+- `[jest-runner]` Correctly serialize `Set` passed to worker ([#9915](https://github.com/facebook/jest/pull/9915))
 
 ### Chore & Maintenance