+++ CHANGELOG.md
@@ -4,11 +4,15 @@
 
 ### Fixes
 
-- `[jest-globals]` Export globals as values, not types ([#9925](https://github.com/facebook/jest/pull/9925))
-
 ### Chore & Maintenance
 
 ### Performance
+
+## 25.5.2
+
+### Fixes
+
+- `[jest-globals]` Export globals as values, not types ([#9925](https://github.com/facebook/jest/pull/9925))
 
 ## 25.5.1